requirements.txt
fastapi==0.115.8
uvicorn==0.15.0
openai==1.60.2
uvloop==0.21.0
duckduckgo-search==7.3.0
pydantic==2.10.6
python-dotenv==0.19.0
httpx==0.27.2
urllib3==1.26.20